Enhance: Image Minifying and Reducing Bundle Size (#35)
* test: try new config for image minifier

* fix: reduce overall bundle size

This includes removing some jquery and adjusting where necessary in the javascript to vanilla js. This is most definitely out of scope for a branch like this and should have been bundled in it's own branch :)

* fix: remove jquery and gifsicle.

What began as a mission to make sure the image minification was working propperly turned into a mission to minify the overall bundle size. Scope creeped, but ultimately this helps the overall goal, and was regned in early enough to not cause crazy delays

* fix: removed console log statments